flush: fix data loss bug
Previously, hits received during a flush (after the /endpoint
request was sent, but before hearing the response) would be thrown
away because reset() was called when the response came back.

Now, we reset the data store right when we serialize it and
under normal circumstances, do nothing. If there's an error
and the config doesn't say we shouldn't, we merge the old
data store with the new one.