package.json: add dev dependencies
And update the Cakefile to match the options used to generate
the file as it is. We dropped the '-b' (--bare) flag to match to
committed compiled file (which *was* wrapped in a function).

Without this change, running `cake build` produced a file that wasn't
wrapped in an IEF (the committed file *is* wrapped).